+++ include/zigzag.h
@@ -1,3 +1,5 @@
 #pragma once
+
+#include <stdint.h>
 // Function to handle inverse zigzag
-void inverse_zigzag(int* vecteur, int **bloc);
+void inverse_zigzag(int* vecteur, int16_t **bloc);
