+++ src/ycbcr_rgb.c
@@ -4,7 +4,7 @@
 #include "ycbcr_rgb.h"
 #include "reader.h"
 
-void YCbCr_to_rgb(uint8_t **Y, uint8_t **Cb, uint8_t **Cr, uint8_t **R, uint8_t **G, uint8_t **B, uint8_t l, uint8_t c) {
+void YCbCr_to_rgb(uint8_t **Y, uint8_t **Cb, uint8_t **Cr, int **R, int **G, int **B, uint8_t l, uint8_t c) {
     for (uint8_t ligne = 0; ligne < l; ligne++) {
         for (uint8_t colonne = 0; colonne < c; colonne++) {
             // Calcul des valeurs RGB
